all was brocken
