+++ js/main.js
@@ -2,7 +2,7 @@ import './data.js';
 import './util.js';
 import {fillCommentsPicture} from './data.js';
 import './miniatures.js';
-import './bigpicture.js';
+import './big-picture.js';
