+++ js/miniatures.js
@@ -1,5 +1,5 @@
 import {photos} from './data.js';
-import {openBigPicture} from './bigpicture.js';
+import {openBigPicture} from './big-picture.js';
 
 const pictureTemplate = document.querySelector('#picture').content;
